still work: Move menu strings to PROGMEM and optimize RAM usage
Refactored menu system to store all menu and item strings in PROGMEM using PGM_P, reducing RAM usage on AVR targets. Updated menu API and implementation to use PGM_P for names and item arrays, and adjusted display rendering to read from PROGMEM. Increased display height, improved display initialization, and increased button debounce time. Old displayWrite function is now commented out.

+++ include/main.h
@@ -22,7 +22,7 @@ extern Adafruit_SSD1306 display;
 // extern INA226 ina;
 
 uint32_t prev_loop_millis = 0;
-float capacity = 0;
+int32_t capacity = 0;
 uint32_t prev_active_curr_millis = 0;
 uint32_t total_active_curr_millis = 0;
 volatile uint16_t voltage = 0;
